+++ rrrc/rrrc_hal.h
@@ -34,8 +34,8 @@ typedef struct
 #define FLASH_HEADER  ((AppFlashHeader_t*) (FLASH_HDR_OFFSET))
 
 void high_res_timer_init(struct timer_descriptor* timer);
-uint32_t high_res_timer_get_count(void);
-uint32_t high_res_timer_ticks_per_ms(void);
+uint16_t high_res_timer_get_count(void);
+uint16_t high_res_timer_ticks_per_ms(void);
 
 static inline uint32_t rtos_get_ticks_in_ms()
 {
